CST150_OC2
This folder has all Activity 14 required deliverables inside. This program features the use of radio buttons, checkboxes, a Listbox as well as the use of labels and buttons. Using this program you're able to submit text from a textbox to the Listbox and from there have the option of saving that text to a file on your Windows computer.